+++ server/src/headers/server.hpp
@@ -3,6 +3,8 @@
 #include "network_session.hpp"
 #include "debug.hpp"
 
+#include <sodium/crypto_box.h>
+#include <sodium/crypto_secretbox.h>
 #include <string>
 #include <unistd.h>
 #include <unordered_map>
@@ -12,6 +14,7 @@
 #include <sys/socket.h>
 #include <arpa/inet.h>
 #include <stdint.h>
+#include <sodium.h>
 
 class Server
 {
@@ -28,6 +31,8 @@ private:
         std::unique_ptr<NetworkSession> session;
         std::string username;
         bool registered{ false };
+        bool key_exchanged{ false };
+        unsigned char client_pk[crypto_box_PUBLICKEYBYTES];
     };
 
     std::unordered_map<int, ClientInfo> m_clients;
@@ -41,11 +46,18 @@ private:
     Debugger m_debugger;
 
     bool m_running{ false };
+
+    unsigned char m_server_pk[crypto_box_PUBLICKEYBYTES];
+    unsigned char m_server_sk[crypto_box_SECRETKEYBYTES];
+    unsigned char m_group_key[crypto_secretbox_KEYBYTES];
+
 private:
     void SetupListener();
     void SetNonBlocking(int fd);
     void EventLoop();
     void HandleNewConnection();
     void HandleClientEvent(int fd, uint32_t events);
-    void BroadcastMessage(const std::string& msg);
+    void BroadcastEncrypted(const std::string& msg);
+    void DisconnectClient(int fd);
+    bool SendSecretbox(NetworkSession* sess, const std::string& msg);
 };